Added the "Explode" transform
This will be useful for the mongo->redshift pipeline. Right now, we
can't create useful "many to many" Foreign Key tables such as school
admin IDs linked to the schools they work with. We can't create such
tables, because we have no way of expanding (or "exploding") the array
of schools associated with the admin object into a set of rows for sql

Other places this transform be useful in bringing data into redshift:
- Coteachers for sections
- Students for sections
- Auth types for launchpads
- Grants for oauthtokens
- Permissions for students/teachers/sections
- Tags for oauthclients